Remove unneeded abstraction in fesvr wrapper
