+++ sim/src/main/cc/endpoints/serial.cc
@@ -1,6 +1,6 @@
 #include "serial.h"
 
-serial_t::serial_t(simif_t* sim, fesvr_proxy_t* fesvr):
+serial_t::serial_t(simif_t* sim, firesim_fesvr_t* fesvr):
     endpoint_t(sim), fesvr(fesvr)
 {
 }
