+++ sim/src/main/cc/firesim/firesim_f1.cc
@@ -4,7 +4,7 @@
 #include "simif_emul.h"
 #endif
 #include "firesim_top.h"
-#include "fesvr/firesim_tsi.h"
+#include "fesvr/firesim_fesvr.h"
 
 #ifdef RTLSIM
 // top for RTL sim
@@ -12,7 +12,7 @@ class firesim_f1_t:
     public simif_emul_t, public firesim_top_t
 {
     public:
-        firesim_f1_t(int argc, char** argv, fesvr_proxy_t* fesvr):
+        firesim_f1_t(int argc, char** argv, firesim_fesvr_t* fesvr):
             firesim_top_t(argc, argv, fesvr) { }
 };
 #else
@@ -21,7 +21,7 @@ class firesim_f1_t:
     public simif_f1_t, public firesim_top_t
 {
     public:
-        firesim_f1_t(int argc, char** argv, fesvr_proxy_t* fesvr):
+        firesim_f1_t(int argc, char** argv, firesim_fesvr_t* fesvr):
             firesim_top_t(argc, argv, fesvr), simif_f1_t(argc, argv) { }
 };
 #endif
@@ -34,8 +34,8 @@ class firesim_f1_t:
 #endif
 
 int main(int argc, char** argv) {
-    firesim_tsi_t tsi(std::vector<std::string>(argv + 1, argv + argc));
-    firesim_f1_t firesim(argc, argv, &tsi);
+    firesim_fesvr_t fesvr(std::vector<std::string>(argv + 1, argv + argc));
+    firesim_f1_t firesim(argc, argv, &fesvr);
     firesim.init(argc, argv);
 
     firesim.run(DESIRED_STEPSIZE);
