+++ sim/src/main/makefrag/firesim/Makefrag
@@ -46,7 +46,7 @@ $(VERILOG) $(HEADER): $(chisel_srcs) $(timestamps)
 driver_dir = $(firesim_base_dir)/src/main/cc
 DRIVER_H = $(shell find $(driver_dir) -name "*.h")
 DRIVER_CC = $(addprefix $(driver_dir)/, $(addsuffix .cc, \
-	firesim/firesim_top fesvr/firesim_tsi fesvr/firesim_fesvr endpoints/serial endpoints/uart \
+	firesim/firesim_top fesvr/firesim_fesvr endpoints/serial endpoints/uart \
 	firesim/firesim_f1 endpoints/simplenic endpoints/blockdev)) $(RISCV)/lib/libfesvr.a
 
 TARGET_CXX_FLAGS := -I$(driver_dir) -I$(driver_dir)/firesim -I$(RISCV)/include
